+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Move root sass file (from `./emma.scss` to `./sass/emma.scss`)
 * Remove snippet: `fw-l`
 * [Bower is deprecated](https://github.com/bower/bower#bower---a-package-manager-for-the-web)
+* Avoid hyphenation at unit value (e.g. `mt-0` -> `mt0`)
 
 ## 0.9.0
 * Restructure based on YAML documents